requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-docx
pandas
tiktoken